Freshclam: Fix issue with CVDCertsDirectory when load-testing databases
Freshclam's feature to load-test a newly downloaded database is not using the
CVDCertsDirectory option to determine the certs directory.
The environment variable works, however.

In addition to the environment variable, you can work around the issue by
disabling load testing.

To fix this issue, we need to extend the libfreshclam API adding a
`fc_test_database_ex()` function that takes the certs directory as a
parameter.

Resolves: https://github.com/Cisco-Talos/clamav/issues/1630

CLAM-2930

+++ libfreshclam/libfreshclam_internal.c
@@ -224,7 +224,7 @@ fc_error_t load_freshclam_dat(void)
 
             size_t minSize = strlen(MIRRORS_DAT_MAGIC) + sizeof(freshclam_dat_v1_t);
             if (minSize > (size_t)file_size) {
-                logg(LOGG_DEBUG, "freshclam.dat is smaller than expected: %zu != %ld\n", sizeof(freshclam_dat_v1_t), file_size);
+                logg(LOGG_DEBUG, "freshclam.dat is smaller than expected: %zu != %lld\n", sizeof(freshclam_dat_v1_t), (long long)file_size);
                 goto done;
             }
